+++ database/seeds/10_RoomStatusSeeder.js
@@ -12,13 +12,16 @@ const Status = use('App/Models/RoomStatus');
 
 class RoomStatusSeeder {
 	async run () {
-		var StatusFiller = ['available', 'deactive', 'maintenance'];
-		var count = await Status.getCount();
+		// load in dummy data for report types
+		let reportStatusFiller = require('../seederData/report_statuses.json');
+
+		// count report types in database
+		const count = await Status.getCount();
 
 		if (count == 0) {
-			for (var i = 0; i < StatusFiller.length; i++) {
+			for (var i = 0; i < reportStatusFiller.length; i++) {
 				const status = new Status();
-				status.name = StatusFiller[i];
+				status.name = reportStatusFiller[i]['name'];
 				await status.save();
 			}
 			console.log('Room Status DB: Finished Seeding');
