+++ database/seeds/6_RoomFeatureSeeder.js
@@ -12,14 +12,9 @@
 
 /** @type {import('@adonisjs/lucid/src/Factory')} */
 const RoomFeature = use('App/Models/RoomFeature');
-const FeaturesRoomsPivot = use('App/Models/FeaturesRoomsPivot');
 
 class RoomFeatureSeeder {
 	async run () {
-
-		// **************************************
-		// SEED ROOM FEATURES
-		// **************************************
 
 		// load in dummy data for room features
 		let featuresFiller = require('../seederData/room_features.json');
@@ -39,28 +34,6 @@ class RoomFeatureSeeder {
 			console.log('Room Features DB: Finished Seeding');
 		} else {
 			console.log('Room Features DB: Already Seeded');
-		}
-
-		// **************************************
-		// SEED FEATURES TO ROOM PIVOT TABLE
-		// **************************************
-
-		// load in dummy data for room features
-		let roomFeaturePivotFillers = require('../seederData/room_features_pivot.json');
-
-		// count room features in database
-		const count = await FeaturesRoomsPivot.getCount();
-
-		if (count == 0) {
-			for (let i = 0; i < featuresFiller.length; i++) {
-				const featuresRoomsPivot = new FeaturesRoomsPivot();
-				featuresRoomsPivot.room_id = roomFeaturePivotFillers[i]['room_id'];
-				featuresRoomsPivot.room_feature_id = roomFeaturePivotFillers[i]['room_feature_id'];
-				await featuresRoomsPivot.save();
-			}
-			console.log('Room Features Pivot DB: Finished Seeding');
-		} else {
-			console.log('Room Features Pivot DB: Already Seeded');
 		}
 	}
 }
